Remove non-decorator code into utils
